+++ packages/lib/gpc/src/index.ts
@@ -1,3 +1,4 @@
+export { type ProtoPODGPCCircuitDesc } from "@pcd/gpcircuits";
 export * from "./gpc";
 export * from "./gpcSerialize";
 export * from "./gpcTypes";
