+++ packages/lib/gpcircuits/package.json
@@ -28,7 +28,7 @@
     "typecheck": "yarn tsc --noEmit",
     "prepublishOnly": "yarn clean && yarn build",
     "test": "ts-mocha --type-check --config ../../../.mocharc.js --exit 'test/**/*.spec.ts'",
-    "clean": "rm -rf dist node_modules *.tsbuildinfo build circuits/test ptau && yarn clean-test-artifacts",
+    "clean": "rm -rf dist node_modules *.tsbuildinfo build circuits/test ptau && yarn clean-prod-artifacts && yarn clean-test-artifacts",
     "gen-circuit-parameters": "tsx scripts/gen-circuit-parameters.ts",
     "gen-test-circuit-parameters": "tsx scripts/gen-circuit-parameters.ts test",
     "gen-artifacts": "tsx scripts/gen-artifacts.ts",
