+++ packages/lib/gpcircuits/scripts/gen-circuit-parameters.ts
@@ -1,11 +1,12 @@
 import { argv } from "process";
-import { ensureCircuitParamSet, jsonFileConfig } from "../src/util";
-import { genCircuitParamConfig } from "./util";
+import { JSON_FILE_CONFIG } from "../src/internal";
+import { ensureCircuitParamSet } from "../src/util";
+import { genCircuitParamConfig } from "./common";
 
 async function main(): Promise<void> {
   const paramType = ensureCircuitParamSet(argv[2]);
   const { circuitParamGenJsonFile, circuitParamJsonFile, circuitsJsonFile } =
-    jsonFileConfig[paramType];
+    JSON_FILE_CONFIG[paramType];
 
   return genCircuitParamConfig(
     circuitParamGenJsonFile,
