+++ packages/pcd/gpc-pcd/src/GPCPCDPackage.ts
@@ -56,14 +56,14 @@ export async function init(args: GPCPCDInitArgs): Promise<void> {
   savedInitArgs = args;
 }
 
-function ensureInitialized(): string {
+function ensureInitialized(): GPCPCDInitArgs {
   if (
     savedInitArgs === undefined ||
     savedInitArgs.zkArtifactPath === undefined
   ) {
     throw new Error("No ZK artifact path given.  Was init skipped?");
   }
-  return savedInitArgs.zkArtifactPath;
+  return savedInitArgs;
 }
 
 async function checkProofArgs(args: GPCPCDArgs): Promise<{
@@ -167,8 +167,7 @@ async function checkProofArgs(args: GPCPCDArgs): Promise<{
  * @throws if the arguments are invalid
  */
 export async function prove(args: GPCPCDArgs): Promise<GPCPCD> {
-  const zkArtifactPath = ensureInitialized();
-  const circuitFamily = savedInitArgs?.circuitFamily;
+  const { zkArtifactPath, circuitFamily } = ensureInitialized();
   const { proofConfig, proofInputs } = await checkProofArgs(args);
   const id =
     args.id !== undefined && typeof args.id.value === "string"
@@ -200,8 +199,7 @@ export async function prove(args: GPCPCDArgs): Promise<GPCPCD> {
  * what was used to prove.
  */
 export async function verify(pcd: GPCPCD): Promise<boolean> {
-  const zkArtifactPath = ensureInitialized();
-  const circuitFamily = savedInitArgs?.circuitFamily;
+  const { zkArtifactPath, circuitFamily } = ensureInitialized();
   return gpcVerify(
     pcd.proof.groth16Proof,
     pcd.claim.config,
