+++ packages/pcd/gpc-pcd/tsconfig.json
@@ -27,9 +27,6 @@
       "path": "../../lib/gpc"
     },
     {
-      "path": "../../lib/gpcircuits"
-    },
-    {
       "path": "../../lib/pcd-types"
     },
     {
